# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Employee Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: getShifts
      tags: [shifts]
      summary: List all shifts
      responses:
        "200":
          description: Shifts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    put:
      operationId: updateShift
      tags: [shifts]
      summary: Update shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        workDays:
          type: string
        lateGraceMinutes:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - startTime
        - endTime
        - workDays
        - lateGraceMinutes
        - createdAt
    CreateShiftRequest:
      type: object
      properties:
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        workDays:
          type: string
        lateGraceMinutes:
          type: integer
      required:
        - name
        - startTime
        - endTime
        - workDays
        - lateGraceMinutes
...